Increase FPS for TinyCraft3D by speeding up trianlge rendering by getting rid of three inner loop divides

+++ src/draw/engine_display_draw.h
@@ -39,6 +39,6 @@ void engine_draw_text(font_resource_class_obj_t *font, mp_obj_t text, float cent
 
 void engine_draw_filled_triangle(uint16_t color, float x0, float y0, float x1, float y1, float x2, float y2, float alpha, engine_shader_t *shader);
 
-void engine_draw_filled_triangle_depth(uint16_t color, int32_t ax, int32_t ay, uint16_t depth_az, int32_t bx, int32_t by, uint16_t depth_bz, int32_t cx, int32_t cy, uint16_t depth_cz, float alpha, engine_shader_t *shader);
+void engine_draw_filled_triangle_depth(uint16_t color, float ax, float ay, uint16_t depth_az, float bx, float by, uint16_t depth_bz, float cx, float cy, uint16_t depth_cz, float alpha, engine_shader_t *shader);
 
 #endif  // ENGINE_DISPLAY_DRAW_H
